libc: Move allocator module into sub-directory allocator/
